we are in the way

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: znajdaou <znajdaou@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/01 20:25:27 by znajdaou          #+#    #+#             */
-/*   Updated: 2024/12/01 20:31:20 by znajdaou         ###   ########.fr       */
+/*   Updated: 2024/12/01 23:06:56 by znajdaou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@
 void				ft_putchar_fd(char c, int fd);
 void				ft_putstr_fd(char *s, int fd);
 void				ft_putnbr_fd(int n, int fd);
-void				ft_putnbr_base(int nbr, char *base);
+void				ft_putnbr_base(long long nbr, char *base);
 
 // ft_printf
 void ft_printf(char *format, ...);
